Make it such that we construct Tree only once (For debug panel)
I've been basically creating+destroying the UI Tree every tick and that's quite costly since we're creating a bunch of smart pointers.

Made a raw pointer to save the state of the Panel. Can have more improvements.

As a result, we do not need to recursively `setBounds` on measure anymore since we're working with the same Tree structure (it'll take a few ticks to put the layout properly, but it's an okay bug to have for now).

I've also passed in a function to `Text` so that it calls the function everytime it wants to render. Ideally, we should memoize it...