Refactor render methods to use Context by value
Now context is passed through the hierarchy and scroll offset is modified and passed through the stack to get visibility of the absolute offset of each element.

As a result, had to modify the `onMouseMove` handler to account for the offset of the scrollPosition.

I think it is absolutely terrible because the handler logic has some counter-intuitive logic. But if it works, it works