revamp positioning of layout from absolute to relative
This makes shit way easier. I used to have a `setBounds` method in `Box` that recursively set the Bounds because I was using absolute positioning for every UI element, so I needed a way to propagate the changes of a parent down.

But if its relative positioning, I can incrementally get information of the eventual absolute positions of each UI element when traversing down the UI element tree.

Stonks!