fix: clip top overflow when scrolling
Discovered that there is a SetClipRectangle function that declares a bounding clip rectangle when drawing onto a surface. That solves the issue of overflow items in the rectangle.

I can now remove the previous complex, buggy way of handling overflow (drawing blanks on the remaining space after drawing a box).